Use TabExpandController in workshop UI
