+++ web/styles/embed/styles.scss
@@ -2,12 +2,12 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-@import "package:mdc_web/material-components-web";
 @import 'package:dart_pad/scss/colors';
+@import 'package:dart_pad/scss/layout';
 @import 'package:dart_pad/scss/shared';
 @import 'package:dart_pad/scss/variables';
 
-@import '../layout';
+@import "package:mdc_web/material-components-web";
 
 * {
   box-sizing: border-box;
