+++ web/styles/styles.scss
@@ -1,8 +1,7 @@
 @import 'package:dart_pad/scss/colors';
-@import 'package:dart_pad/scss/variables';
+@import 'package:dart_pad/scss/layout';
 @import 'package:dart_pad/scss/shared';
-
-@import 'layout';
+@import 'package:dart_pad/scss/variables';
 
 // Material Design Web theme colors. Must be imported before importing
 // material-components-web.scss.
@@ -259,67 +258,6 @@ a {
   outline: 0 solid transparent;
 }
 
-// Editor Tabs
-.editor-tab-host {
-  @include layout-vertical;
-}
-
-.border-top {
-  border-top: 6px solid $dark-gutter-background-color !important;
-}
-
-.editor-tabs {
-  @include layout-horizontal;
-  @include layout-center;
-  @include layout-justified;
-  padding-left: 8px;
-
-  button {
-    margin-right: 8px;
-  }
-  .tab-group {
-    @include layout-horizontal;
-    @include layout-center;
-  }
-}
-
-.editor-tab {
-  @include mdc-button-ink-color(#f8f9fa);
-  text-transform: none !important;
-  letter-spacing: normal;
-
-  color: $label-color;
-  border-top: 2px solid transparent;
-  padding: 8px 4px;
-
-  &.active {
-    color: $dark-blue;
-    font-weight: bold;
-  }
-}
-
-.mdc-tab__content {
-  text-transform: none !important;
-}
-
-.mdc-tab--active {
-  .mdc-tab__text-label {
-    font-weight: normal;
-  }
-}
-
-#editor-panel-close-button {
-  @include mdc-icon-button-size(16px, 16px, 8px);
-}
-
-#editor-panel-tab-host {
-  @include layout-flex;
-  @include layout-vertical;
-  .documentation {
-    @include layout-flex;
-  }
-}
-
 // documentation and parameter info styling
 .documentation {
   padding: $doc-console-padding;
@@ -495,6 +433,23 @@ a {
     height: auto;
     width: 20px;
     margin-right: 14px;
+}
+
+// Editor Tabs
+.editor-tab {
+  @include mdc-button-ink-color(#f8f9fa);
+  color: $label-color;
+  &.active {
+    color: $dark-blue;
+  }
+}
+
+.border-top {
+  border-top: 6px solid $dark-gutter-background-color !important;
+}
+
+#editor-panel-close-button {
+  @include mdc-icon-button-size(16px, 16px, 8px);
 }
 
 // responsive layout
